fix: Renaming a nonexistent file will report an error
单元检测的日志文件不存在不要重命名，存在则重命名，不存在重命名会报错，没有检测到内存泄露的时候就会不存在。

Log:

+++ tests/test-recoverage-qmake.sh
@@ -27,4 +27,4 @@ genhtml -o $REPORT_DIR $BUILD_DIR/coverage.info
 #rm -rf $BUILD_DIR
 #rm -rf ../$BUILD_DIR
 
-mv ./build/asan.log* ./build/asan_dtkcore.log
+test -e ./build/asan.log* && mv ./build/asan.log* ./build/asan_dtkcore.log
